# api/openapi_scib_inception_phase_demo.yaml
openapi: 3.0.3
info:
  title: scib_inception_phase_demo API
  version: 1.0.0
  description: API specification for task management system with comprehensive error handling
servers:
  - url: https://example.com
    description: Production server
paths:
  /api/tasks:
    post:
      operationId: DEMO-760CreateTaskPOST
      summary: Create a new task
      description: Creates a new task with error handling for database connection failures, timeouts, and unexpected server errors
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '408':
          description: Request timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tasks/{id}:
    get:
      operationId: DEMO-760GetTaskGET
      summary: Get task by ID
      description: Retrieves a task by its unique identifier
      tags:
        - Tasks
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '408':
          description: Request timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: DEMO-760UpdateTaskPUT
      summary: Update task
      description: Updates an existing task with new information
      tags:
        - Tasks
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '408':
          description: Request timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: DEMO-760DeleteTaskDELETE
      summary: Delete task
      description: Deletes a task by its unique identifier
      tags:
        - Tasks
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '408':
          description: Request timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    TaskCreateRequest:
      type: object
      required:
        - title
        - description
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title
          example: Complete project documentation
        description:
          type: string
          minLength: 1
          maxLength: 1000
          description: Task description
          example: Write comprehensive documentation for the task management system
    TaskUpdateRequest:
      type: object
      required:
        - title
        - description
        - status
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title
          example: Complete project documentation
        description:
          type: string
          minLength: 1
          maxLength: 1000
          description: Task description
          example: Write comprehensive documentation for the task management system
        status:
          type: string
          enum:
            - PENDING
            - IN_PROGRESS
            - COMPLETED
            - CANCELLED
          description: Task status
          example: IN_PROGRESS
    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Task unique identifier
          example: 1
        title:
          type: string
          description: Task title
          example: Complete project documentation
        description:
          type: string
          description: Task description
          example: Write comprehensive documentation for the task management system
        status:
          type: string
          enum:
            - PENDING
            - IN_PROGRESS
            - COMPLETED
            - CANCELLED
          description: Task status
          example: PENDING
        createdAt:
          type: string
          format: date-time
          description: Task creation timestamp
          example: 2024-01-15T10:30:00Z
        updatedAt:
          type: string
          format: date-time
          description: Task last update timestamp
          example: 2024-01-15T14:45:00Z
    ErrorResponse:
      type: object
      required:
        - errorCode
        - message
        - timestamp
      properties:
        errorCode:
          type: string
          description: Error code identifier
          example: SERVICE_UNAVAILABLE
          enum:
            - SERVICE_UNAVAILABLE
            - REQUEST_TIMEOUT
            - VALIDATION_ERROR
            - INTERNAL_SERVER_ERROR
        message:
          type: string
          description: Human-readable error message
          example: Service temporarily unavailable, please try again later
        timestamp:
          type: integer
          format: int64
          description: Error timestamp in milliseconds
          example: 1705319400000
        details:
          type: array
          description: Additional error details
          items:
            type: string
          example:
            - Database connection failed
            - Retry attempt 1 of 3
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          description: Success message
          example: Operation completed successfully
        timestamp:
          type: integer
          format: int64
          description: Response timestamp in milliseconds
          example: 1705319400000
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT-based authentication
security:
  - bearerAuth: []